+++ asn1_group/B167_cb_dcall.c
@@ -4,7 +4,7 @@
 #include "LTE-ML1-Random-Access-Request-MSG1-Report-S-H.h"
 #include "B167_cb_dcall.h"
 
-void convert_S_H_B167(uint8_t *hex_data, size_t *index)
+void convert_S_H_B167(uint8_t *hex_data, int *index)
 {
     //version and skip0
     *index+=4;
@@ -30,7 +30,7 @@ void convert_S_H_B167(uint8_t *hex_data, size_t *index)
     *index+=2;
 }
 
-void decode_B167(uint8_t *hex_data, size_t length, size_t *index)
+void decode_B167(uint8_t *hex_data, size_t length, int *index)
 {
     LTE_ML1_Random_Access_Request_MSG1_Report_S_H_t  *t_S_H = 0;
     asn_dec_rval_t rval_S_H;
